* save beer edits * change beer percentage to bigdecimal * make detail texts as edittext * make edit text clickthrough when disabled

+++ app/src/main/java/com/github/vmas/beermeter/core/model/Beer.kt
@@ -2,6 +2,7 @@ package com.github.vmas.beermeter.core.model
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
+import java.math.BigDecimal
 
 /**
  * Created by Sławomir Golonka @ ConciseSoftware on 06-02-2020.
@@ -11,7 +12,7 @@ data class Beer(
     val name: String,
     val imgUrl: String?,
     val type: String,
-    val percentage: Float,
+    val percentage: BigDecimal,
     val country: String,
     val website: String?
 ) : Parcelable
